+++ projections/typescript/index.ts
@@ -1,3 +1,4 @@
 export * from "./capsule.js";
+export * from "./validator-envelope-families.js";
 export * from "./validator-routes.js";
 export * from "./validator-api.js";